requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
attrs==26.1.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==47.0.0
decorator==5.2.1
distro==1.9.0
dotenv==0.9.9
executing==2.2.1
frozenlist==1.8.0
google-auth==2.49.2
google-genai==1.73.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
ipython==8.39.0
jedi==0.19.2
matplotlib-inline==0.2.1
multidict==6.7.1
openai==0.27.10
parso==0.8.6
pexpect==4.9.0
prompt_toolkit==3.0.52
propcache==0.4.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
python-dotenv==1.2.2
regex==2026.4.4
requests==2.33.1
sniffio==1.3.1
stack-data==0.6.3
tabulate==0.9.0
tenacity==9.1.4
tiktoken==0.3.3
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
websockets==16.0
yarl==1.23.0
